"Add email input and sign up button to last section; update CSS for new elements"
This commit adds a new div containing an email input field and a sign up button to the last section of the HTML document. It also updates the CSS to style these new elements and adjust the layout of the last section. The JavaScript code is also updated to select the new elements using the "nth-of-type" selector instead of "last-of-type".